change:  - cleanup consoles  - move feeswitcher change to hook  - add drawer for switch fees for mobile view  - disable feeswitcher on nav in mobile view
