+++ packages/boba/gateway/src/components/layout/Header/feeSwitcher/styles.ts
@@ -1,12 +1,16 @@
 import { HelpOutline } from '@mui/icons-material'
 import { Heading, Typography } from 'components/global'
-import styled from 'styled-components'
+import styled, { css } from 'styled-components'
+import { mobile } from 'themes/screens'
 
 export const FeeSwitcherWrapper = styled.div`
   display: flex;
   gap: 5px;
   align-items: center;
   justify-content: space-around;
+  ${mobile(css`
+    display: none;
+  `)}
 `
 
 export const FeeSwitcherLabel = styled(Heading).attrs({
