+++ packages/boba/gateway/src/hooks/useWalletConnect.ts
@@ -44,7 +44,6 @@ export const useWalletConnect = () => {
    */
 
   const triggerInit = useCallback(() => {
-    console.log(['network', network, networkType])
     const initAccount = async () => {
       const initialized = await networkService.initializeAccount({
         chainIdChanged,
